# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ObtainResponseErrorJSON": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"details": {}
				},
				"required": [
					"name",
					"message",
					"details"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NLUResponse": {
				"description": "The response object to the client.",
				"properties": {
					"intent": {
						"type": "string",
						"description": "A message utterance is mapped to an intent",
						"example": "clock"
					},
					"confidence": {
						"type": "number",
						"format": "double",
						"description": "Confidence level for the intent (%)",
						"example": "95"
					}
				},
				"required": [
					"intent",
					"confidence"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "test-service",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/best_response": {
			"get": {
				"operationId": "GetBestResponse",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NLUResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad request"
					},
					"422": {
						"description": "Obtain NLUResponse Failure",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ObtainResponseErrorJSON"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "Obtain Response Error",
											"message": "Operation failed",
											"details": {
												"requestBody": {
													"message": "cannot get a suitable response from server",
													"value": "message_text"
												}
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Returns the best suitable answer to a client message. The answer will be obtained\nasking to the different NLU servers.",
				"security": [],
				"parameters": [
					{
						"description": "The client message that needs a server answer.",
						"in": "query",
						"name": "message",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}